Remove last traces of old load balancer
