+++ meta1v2/meta1_backend.h
@@ -1,7 +1,7 @@
 /*
 OpenIO SDS meta1v2
 Copyright (C) 2014 Worldine, original work as part of Redcurrant
-Copyright (C) 2015 OpenIO, modified as part of OpenIO Software Defined Storage
+Copyright (C) 2015-2016 OpenIO, as part of OpenIO Software Defined Storage
 
 This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU Affero General Public License as
@@ -69,11 +69,9 @@ enum m1v2_getsrv_e
 	M1V2_GETSRV_DRYRUN = 0x02,
 };
 
-struct grid_lb_iterator_s;
 struct meta1_prefixes_set_s;
 struct sqlx_repository_s;
 struct sqlx_sqlite3_s;
-struct grid_lbpool_s;
 struct oio_url_s;
 
 struct meta1_backend_s;
