+++ meta2v2/meta2_utils.h
@@ -1,7 +1,7 @@
 /*
 OpenIO SDS meta2v2
 Copyright (C) 2014 Worldine, original work as part of Redcurrant
-Copyright (C) 2015 OpenIO, modified as part of OpenIO Software Defined Storage
+Copyright (C) 2015-2016 OpenIO, as part of OpenIO Software Defined Storage
 
 This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU Affero General Public License as
@@ -38,8 +38,6 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 struct storage_policy_s;
 struct oio_url_s;
-struct grid_lb_iterator_s;
-struct lb_next_opt_s;
 struct sqlx_sqlite3_s;
 
 struct list_params_s
